# src/main/resources/swagger/example/openapi.json
{
  "openapi": "3.0.0",
  "servers": [
    {
      "description": "Development server",
      "url": "http://localhost:8080/api"
    }
  ],
  "info": {
    "description": "This is a simple example API for Helpdesk",
    "version": "1.0.0",
    "title": "Simple Helpdesk API",
    "contact": {
      "email": "mentor@novahub.vn"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
...
  "paths": {
...
    "/skills": {
      "get": {
        "tags": [
          "Skill"
        ],
        "summary": "Get all skills",
        "description": "Get all skills",
        "parameters": [
          {
            "in": "header",
            "name": "access_token",
            "required": true,
            "schema": {
              "type": "string",
              "example": "dff4d0d4891bfd8bafdb859dac4cbccb218bcf5adc9887a627c6c159f76ef0b9"
            }
          },
          {
            "in": "query",
            "name": "categoryId",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            }
          },
          {
            "in": "query",
            "name": "keyword",
            "schema": {
              "type": "string",
              "example": "Java"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "type": "string",
              "enum": [
                "id,asc",
                "id,desc",
                "name,asc",
                "name,desc",
                "createdAt,asc",
                "createdAt,desc",
                "updatedAt,asc",
                "updatedAt,desc"
              ],
              "example": "name,desc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get all skills successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Sort",
                  "properties": {
                    "content": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Skill",
                        "properties": {
                          "category": {
                            "$ref": "#/components/schemas/Category"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Token is expired",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "timestamp": {
                      "type": "string",
                      "example": 1532589762.947
                    },
                    "status": {
                      "type": "integer",
                      "example": 400
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Token is expired"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "TokenIsExpiredException with token = 579989c508208d49c978e6cb99383688e6f1fedabfc22c27d3542ab0341d3b13"
                    },
                    "path": {
                      "type": "string",
                      "example": "/skills"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "timestamp": {
                      "type": "string",
                      "example": "1532506012.933"
                    },
                    "status": {
                      "type": "integer",
                      "format": "int64",
                      "example": 401
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Unauthorized"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid token"
                    },
                    "path": {
                      "type": "string",
                      "example": "/skills"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Access is denied",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "timestamp": {
                      "type": "string",
                      "example": "1532506012.933"
                    },
                    "status": {
                      "type": "integer",
                      "format": "int64",
                      "example": 403
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Access is denied"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Access is denied"
                    },
                    "path": {
                      "type": "string",
                      "example": "/skills"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/skills/{id}": {
      "get": {
        "tags": [
          "Skill"
        ],
        "summary": "Get a skill by id",
        "description": "Get a skill by id",
        "parameters": [
          {
            "in": "header",
            "name": "access_token",
            "required": true,
            "schema": {
              "type": "string",
              "example": "dff4d0d4891bfd8bafdb859dac4cbccb218bcf5adc9887a627c6c159f76ef0b9"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Skill",
                  "properties": {
                    "category": {
                      "$ref": "#/components/schemas/Category"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Token is expired",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "timestamp": {
                      "type": "string",
                      "example": 1532589762.947
                    },
                    "status": {
                      "type": "integer",
                      "example": 400
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Token is expired"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "TokenIsExpiredException with token = 579989c508208d49c978e6cb99383688e6f1fedabfc22c27d3542ab0341d3b13"
                    },
                    "path": {
                      "type": "string",
                      "example": "/skills/{id}"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "timestamp": {
                      "type": "string",
                      "example": "1532506012.933"
                    },
                    "status": {
                      "type": "integer",
                      "format": "int64",
                      "example": 401
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Unauthorized"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid token"
                    },
                    "path": {
                      "type": "string",
                      "example": "/skills/{id}"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Access is denied",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "timestamp": {
                      "type": "string",
                      "example": "1532506012.933"
                    },
                    "status": {
                      "type": "integer",
                      "format": "int64",
                      "example": 403
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Access is denied"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Access is denied"
                    },
                    "path": {
                      "type": "string",
                      "example": "/skills/{id}"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "skill not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "timestamp": {
                      "type": "string",
                      "example": "1532462094.006"
                    },
                    "status": {
                      "type": "integer",
                      "example": 404
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Skill not found"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "SkillNotFoundException with id = {id}"
                    },
                    "path": {
                      "type": "string",
                      "example": "/skills/{id}"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/skills/{id}/users": {
      "get": {
        "tags": [
          "Skill"
        ],
        "summary": "Get a skill by id",
        "description": "Get a skill by id",
        "parameters": [
          {
            "in": "header",
            "name": "access_token",
            "required": true,
            "schema": {
              "type": "string",
              "example": "dff4d0d4891bfd8bafdb859dac4cbccb218bcf5adc9887a627c6c159f76ef0b9"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "type": "string",
              "enum": [
                "id,asc",
                "id,desc",
                "email,asc",
                "email,desc",
                "status,asc",
                "status,desc",
                "firstName,asc",
                "firstName,desc",
                "lastName,asc",
                "lastName,desc",
                "address,asc",
                "address,desc"
              ],
              "example": "firstName,desc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Sort",
                  "properties": {
                    "content": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Account",
                        "properties": {
                          "role": {
                            "$ref": "#/components/schemas/Role"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Token is expired",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "timestamp": {
                      "type": "string",
                      "example": 1532589762.947
                    },
                    "status": {
                      "type": "integer",
                      "example": 400
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Token is expired"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "TokenIsExpiredException with token = 579989c508208d49c978e6cb99383688e6f1fedabfc22c27d3542ab0341d3b13"
                    },
                    "path": {
                      "type": "string",
                      "example": "/skills/{id}/users"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "timestamp": {
                      "type": "string",
                      "example": "1532506012.933"
                    },
                    "status": {
                      "type": "integer",
                      "format": "int64",
                      "example": 401
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Unauthorized"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid token"
                    },
                    "path": {
                      "type": "string",
                      "example": "/skills/{id}/users"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Access is denied",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "timestamp": {
                      "type": "string",
                      "example": "1532506012.933"
                    },
                    "status": {
                      "type": "integer",
                      "format": "int64",
                      "example": 403
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Access is denied"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Access is denied"
                    },
                    "path": {
                      "type": "string",
                      "example": "/skills/{id}/users"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "skill not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "timestamp": {
                      "type": "string",
                      "example": "1532462094.006"
                    },
                    "status": {
                      "type": "integer",
                      "example": 404
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Skill not found"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "SkillNotFoundException with id = 1"
                    },
                    "path": {
                      "type": "string",
                      "example": "/skills/1"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Account": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "email": {
            "type": "string",
            "example": "helpdesk@novahub.vn"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "LOCKED"
            ],
            "example": "ACTIVE"
          },
          "first_name": {
            "type": "string",
            "example": "help"
          },
          "last_name": {
            "type": "string",
            "example": "desk"
          },
          "birth_day": {
            "type": "string",
            "format": "date",
            "description": "pattern yyyy-MM-dd",
            "example": "1997-09-01T00:00:00.000Z"
          },
          "address": {
            "type": "string",
            "example": "14 An Thuong 32 street, Da Nang city"
          },
          "phone": {
            "type": "string",
            "example": 20725
          },
          "title": {
            "type": "string",
            "example": "Junior Front-end"
          },
          "introduction": {
            "type": "string",
            "example": "My name is Than Quang Minh. I come from Vietnam. Currently I work as a web developer. I have more than 15 years programming and more than 5 years of work experience in Software & web development using .NET, PHP, RoR, Java Spring Framework and adequate tutoring & on-job training experience, as well as strong background in Computer Science, computer system, database modeling & information architecture, and fluent in algorithmic & OOP programming and web-related technologies."
          },
          "avatar_url": {
            "type": "string",
            "example": "http://localhost:8080/avatar.png"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2018-07-24 17:50:45"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "example": "2018-07-24 17:50:45"
          },
          "role_id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          }
        }
      },
      "Role": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "name": {
            "type": "string",
            "enum": [
              "ADMIN",
              "CLERK",
              "EMPLOYEE"
            ],
            "example": "ADMIN"
          }
        }
      },
...
      "Skill": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Java"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2018-07-24 18:39:48"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "example": "2018-07-24 18:39:48"
          },
          "category_id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          }
        }
      },
...
      "Category": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Back-end"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2018-07-24 17:50:45"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "example": "2018-07-24 17:50:45"
          }
        }
      },
...
      "Sort": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "pageable": {
            "type": "object",
            "properties": {
              "sort": {
                "type": "object",
                "properties": {
                  "sorted": {
                    "type": "boolean",
                    "example": false
                  },
                  "unsorted": {
                    "type": "boolean",
                    "example": true
                  }
                }
              },
              "pageSize": {
                "type": "integer",
                "format": "int64",
                "example": 20
              },
              "pageNumber": {
                "type": "integer",
                "format": "int64",
                "example": 0
              },
              "offset": {
                "type": "integer",
                "format": "int64",
                "example": 0
              },
              "paged": {
                "type": "boolean",
                "example": true
              },
              "unpaged": {
                "type": "boolean",
                "example": false
              }
            }
          },
          "last": {
            "type": "boolean",
            "example": true
          },
          "first": {
            "type": "boolean",
            "example": true
          },
          "totalPages": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "totalElements": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "sort": {
            "type": "object",
            "properties": {
              "sorted": {
                "type": "boolean",
                "example": false
              },
              "unsorted": {
                "type": "boolean",
                "example": true
              }
            }
          },
          "size": {
            "type": "integer",
            "format": "int64",
            "example": 20
          },
          "number": {
            "type": "integer",
            "format": "int64",
            "example": 0
          }
        }
      }
    }
  }
}